test(jazz): organize row policy coverage
