+++ crates/jazz/src/node/tests/policies_rls/mod.rs
@@ -1,0 +1,10 @@
+// Read- and write-policy behavior grouped by the authorization boundary under test.
+
+include!("support.rs");
+include!("write_authorization.rs");
+include!("read_delivery.rs");
+include!("includes.rs");
+include!("maintained_views.rs");
+include!("fail_closed.rs");
+include!("recursive.rs");
+include!("windows.rs");
